Update to latest version

+++ acc_share_del.php
@@ -1,16 +1,17 @@
 <!DOCTYPE HTML>
 <?PHP
 	require 'functions.php';
-	check_logon();
-	check_delete();
+	checkLogin();
+	checkPermissionDelete();
 	connect();
 
 	//Delete from SHARES
 	if (isset($_GET['sha_id'])){
 		$sha_id = sanitize($_GET['sha_id']);
+		
 		$sql_delsav = "DELETE FROM shares WHERE share_id = $sha_id";
 		$query_delsav = mysql_query($sql_delsav);
-		check_sql($query_delsav);
+		checkSQL($query_delsav);
 	}
 	
 	//Refer to customer.php
